+++ forge/src/main/java/com/christofmeg/justenoughbreeding/emi/BreedingCategoryEMI.java
@@ -65,16 +65,15 @@ public class BreedingCategoryEMI extends AbstractRecipeCategoryEMI {
         widgets.addSlot(EmiIngredient.of(recipe.spawnEgg), 149, 1);
 
         int inputX = 69 + 5;
-        int inputY = 58 - 10;
-        int extraY = inputY - 19;
+        int inputY = 48;
         boolean hasExtraInput = recipe.extraInputStack != null && !recipe.extraInputStack.isEmpty();
         boolean hasOutput = recipe.resultItemStack != null && !recipe.resultItemStack.isEmpty();
         if (hasExtraInput && hasOutput) {
-            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX, inputY + 9);
-            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), inputX, extraY + 9);
+            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX, 38);
+            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), inputX, 57);
         } else if (hasExtraInput) {
-            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX + 33, inputY + 9);
-            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), inputX + 33, extraY + 9);
+            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX + 33, 38);
+            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), inputX + 33, 57);
         } else if (!hasOutput) {
             widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX  + 33, inputY);
         } else {