+++ forge/src/main/java/com/christofmeg/justenoughbreeding/rei/TemperCategoryREI.java
@@ -40,8 +40,8 @@ public class TemperCategoryREI extends AbstractRecipeCategoryREI<TemperDisplay> 
                 !display.getExtraInputEntries().get(0).get(0).isEmpty();
 
         if (hasExtraInput) {
-            widgets.add(Widgets.createSlot(new Point(bounds.getX() + 112, bounds.getCenterY() + 12)).entries(display.getInputEntries().get(0)));
-            widgets.add(Widgets.createSlot(new Point(bounds.getX() + 112, bounds.getCenterY() - 11)).entries(display.getExtraInputEntries().get(0)));
+            widgets.add(Widgets.createSlot(new Point(bounds.getX() + 112, bounds.getCenterY() - 11)).entries(display.getInputEntries().get(0)));
+            widgets.add(Widgets.createSlot(new Point(bounds.getX() + 112, bounds.getCenterY() + 12)).entries(display.getExtraInputEntries().get(0)));
         } else {
             widgets.add(Widgets.createSlot(new Point(bounds.getX() + 112, bounds.getCenterY() + 3)).entries(display.getInputEntries().get(0)));
         }